+++ build-systems/obs/cmake-qt6/build-for-obs.sh
@@ -40,7 +40,7 @@ cd QOwnNotes || exit 1
 git submodule update --init
 
 # build binary translation files
-lrelease src/QOwnNotes.pro
+/usr/lib/qt6/bin/lrelease src/QOwnNotes.pro
 
 # remove huge .git folder
 rm -Rf .git